requirements.txt
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
